+++ vec.h
@@ -15,7 +15,7 @@
 
 #define VEC_STACK_VEC(X, SZ)                                                                                           \
     auto X##_size = (SZ);                                                                                              \
-    double* X##_storage = (double*)_alloca(sizeof(double) * X##_size);                                                 \
+    float* X##_storage = (float*)_alloca(sizeof(float) * X##_size);                                                 \
     vec_slice X(X##_storage, X##_size)
 
 struct mat_slice;
@@ -23,20 +23,20 @@ struct mat_slice;
 struct vec_slice
 {
     constexpr vec_slice() = default;
-    constexpr vec_slice(double* data, size_t len) : m_data(data), m_len(len) { }
+    constexpr vec_slice(float* data, size_t len) : m_data(data), m_len(len) { }
     template<size_t Sz>
-    constexpr vec_slice(double (&data)[Sz]) : m_data(data), m_len(Sz)
+    constexpr vec_slice(float (&data)[Sz]) : m_data(data), m_len(Sz)
     {
     }
-    vec_slice(std::valarray<double>& v) : m_data(&v[0]), m_len(v.size()) { }
+    vec_slice(std::valarray<float>& v) : m_data(&v[0]), m_len(v.size()) { }
 
-    constexpr double* data() { return m_data; }
+    constexpr float* data() { return m_data; }
     constexpr size_t size() const { return m_len; }
 
-    double dot(vec_slice o) const
+    float dot(vec_slice o) const
     {
         VEC_CHECK_BOUNDS(o);
-        double sum = 0.0;
+        float sum = 0.0;
         for (size_t i = 0; i < m_len; ++i)
         {
             sum += m_data[i] * o.m_data[i];
@@ -44,7 +44,7 @@ struct vec_slice
         return sum;
     }
 
-    double& operator[](size_t i)
+    float& operator[](size_t i)
     {
 #if !defined(NDEBUG) || defined(VEC_ENABLE_CHECKS)
         if (i >= m_len) std::terminate();
@@ -52,8 +52,8 @@ struct vec_slice
         return m_data[i];
     }
 
-    double* begin() { return m_data; }
-    double* end() { return m_data + m_len; }
+    float* begin() { return m_data; }
+    float* end() { return m_data + m_len; }
 
     vec_slice slice() { return *this; }
     vec_slice slice(size_t offset) { return {m_data + offset, m_len - offset}; }
@@ -77,7 +77,7 @@ struct vec_slice
         VEC_CHECK_BOUNDS(o);
         VEC_EOP(= o.m_data[i]);
     }
-    vec_slice& assign(double d) { VEC_EOP(= d); }
+    vec_slice& assign(float d) { VEC_EOP(= d); }
 
     vec_slice& assign_mult(vec_slice a, vec_slice b)
     {
@@ -85,7 +85,7 @@ struct vec_slice
         VEC_CHECK_BOUNDS(b);
         VEC_EOP(= a.m_data[i] * b.m_data[i]);
     }
-    vec_slice& assign_mult(vec_slice a, double b)
+    vec_slice& assign_mult(vec_slice a, float b)
     {
         VEC_CHECK_BOUNDS(a);
         VEC_EOP(= a.m_data[i] * b);
@@ -97,7 +97,7 @@ struct vec_slice
         VEC_CHECK_BOUNDS(b);
         VEC_EOP(= a.m_data[i] + b.m_data[i]);
     }
-    vec_slice& assign_add(vec_slice a, double b)
+    vec_slice& assign_add(vec_slice a, float b)
     {
         VEC_CHECK_BOUNDS(a);
         VEC_EOP(= a.m_data[i] + b);
@@ -109,7 +109,7 @@ struct vec_slice
         VEC_CHECK_BOUNDS(b);
         VEC_EOP(= a.m_data[i] - b.m_data[i]);
     }
-    vec_slice& assign_sub(double a, vec_slice b)
+    vec_slice& assign_sub(float a, vec_slice b)
     {
         VEC_CHECK_BOUNDS(b);
         VEC_EOP(= a - b.m_data[i]);
@@ -121,7 +121,7 @@ struct vec_slice
         VEC_CHECK_BOUNDS(b);
         VEC_EOP(+= a.m_data[i] * b.m_data[i]);
     }
-    vec_slice& fma(vec_slice a, double b)
+    vec_slice& fma(vec_slice a, float b)
     {
         VEC_CHECK_BOUNDS(a);
         VEC_EOP(+= a.m_data[i] * b);
@@ -133,9 +133,9 @@ struct vec_slice
         VEC_EOP(+= a.m_data[i]);
     }
 
-    vec_slice& mult(double a) { VEC_EOP(*= a); }
+    vec_slice& mult(float a) { VEC_EOP(*= a); }
 
-    double max(double init) const
+    float max(float init) const
     {
         for (size_t i = 0; i < m_len; ++i)
         {
@@ -145,14 +145,14 @@ struct vec_slice
     }
 
     // ratio is multiplied by x
-    vec_slice& decay_average(vec_slice x, double ratio)
+    vec_slice& decay_average(vec_slice x, float ratio)
     {
         VEC_CHECK_BOUNDS(x);
         VEC_EOP(= m_data[i] * (1 - ratio) + x.m_data[i] * ratio);
     }
 
     // ratio is multiplied by x
-    vec_slice& decay_variance(vec_slice x, double ratio)
+    vec_slice& decay_variance(vec_slice x, float ratio)
     {
         VEC_CHECK_BOUNDS(x);
         VEC_EOP(= m_data[i] * (1 - ratio) + x.m_data[i] * x.m_data[i] * ratio);
@@ -161,14 +161,14 @@ struct vec_slice
 #undef VEC_EOP
 
 private:
-    double* m_data = nullptr;
+    float* m_data = nullptr;
     size_t m_len = 0;
 };
 
 struct mat_slice
 {
     constexpr mat_slice() = default;
-    constexpr mat_slice(double* data, size_t rows, size_t cols) : m_data(data), m_rows(rows), m_cols(cols) { }
+    constexpr mat_slice(float* data, size_t rows, size_t cols) : m_data(data), m_rows(rows), m_cols(cols) { }
     constexpr mat_slice(vec_slice data, size_t stride)
         : m_data(data.data()), m_rows(data.size() / stride), m_cols(stride)
     {
@@ -177,7 +177,7 @@ struct mat_slice
 #endif
     }
 
-    double* data() { return m_data; }
+    float* data() { return m_data; }
     size_t size() const { return m_rows * m_cols; }
     size_t rows() const { return m_rows; }
     size_t cols() const { return m_cols; }
@@ -198,13 +198,13 @@ struct mat_slice
         return vec_slice(m_data + (m_rows - 1) * m_cols, m_cols);
     }
 
-    double* begin() { return m_data; }
-    double* end() { return m_data + size(); }
+    float* begin() { return m_data; }
+    float* end() { return m_data + size(); }
 
     vec_slice flat() { return {data(), size()}; }
 
 private:
-    double* m_data = nullptr;
+    float* m_data = nullptr;
     size_t m_rows = 0;
     size_t m_cols = 0;
 };
@@ -229,10 +229,10 @@ struct vec
     vec_slice slice(size_t offset) { return {m_data.get() + offset, m_len - offset}; }
     vec_slice slice(size_t offset, size_t len) { return {m_data.get() + offset, len}; }
 
-    double* data() { return m_data.get(); }
+    float* data() { return m_data.get(); }
     size_t size() const { return m_len; }
 
-    double& operator[](size_t i)
+    float& operator[](size_t i)
     {
 #if !defined(NDEBUG) || defined(VEC_ENABLE_CHECKS)
         if (i >= m_len) std::terminate();
@@ -240,13 +240,13 @@ struct vec
         return m_data[i];
     }
 
-    double* begin() { return m_data.get(); }
-    double* end() { return m_data.get() + m_len; }
+    float* begin() { return m_data.get(); }
+    float* end() { return m_data.get() + m_len; }
 
-    const double* begin() const { return m_data.get(); }
-    const double* end() const { return m_data.get() + m_len; }
+    const float* begin() const { return m_data.get(); }
+    const float* end() const { return m_data.get() + m_len; }
 
-    void realloc(size_t n, double v)
+    void realloc(size_t n, float v)
     {
         realloc_uninitialized(n);
         slice().assign(v);
@@ -261,7 +261,7 @@ struct vec
         }
         else if (m_len != n)
         {
-            m_data = std::make_unique<double[]>(n);
+            m_data = std::make_unique<float[]>(n);
             m_len = n;
         }
     }
@@ -273,7 +273,7 @@ struct vec
     }
 
 private:
-    std::unique_ptr<double[]> m_data;
+    std::unique_ptr<float[]> m_data;
     size_t m_len = 0;
 };
